Refactor Kafka tweet producer
